fixed schedule glitched
