+++ client/src/HomePage/schedule/TotalUnits.tsx
@@ -3,7 +3,6 @@ import {RootState} from '../../app/store';
 
 const TotalUnits = () => {
     const units = useSelector((state:RootState)=> state.store.totalUnits);
-
     return (
       <div>
         <p> Total Units:</p>
